refactor: 회원가입 버튼 섹션 모듈화
