+++ src/app/(without-side)/signup/page.tsx
@@ -4,22 +4,18 @@ import Checkbox from '@/components/features/signup/checkbox';
 import Image from 'next/image';
 import { useEffect, useState } from 'react';
 import Nickname from '@/components/common/account/nickname';
-import { UpdatePrivacyRequest, UpdateUserInfoRequest } from '@/services/apis/user/dto';
-import { updatePrivacy, updateUserInfo } from '@/services/apis/user';
 import { agreementDatas } from '@/lib/constants/agreementDatas';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { getCookie, setCookie } from '@/lib/utils';
-import { DOMAIN_URL, SERVER_URL } from '@/services/config/constants';
 import FavoriteTeamSection from '@/components/common/account/favorite-team-section';
-import BottomButton from '@/components/common/bottom-button';
+import ButtonSection from '@/components/features/signup/button-section';
+
 export default function Page() {
 	const router = useRouter();
 	const searchParams = useSearchParams();
 	const provider = searchParams.get('provider');
 	const socialLogoUrl = provider === 'naver' ? '/sns/naver-small.svg' : '/sns/kakao-small.svg';
 	const socialLogoAlt = provider === 'naver' ? '네이버 로고 이미지' : '카카오 로고 이미지';
-
-	const [isValidAccess, setIsValidAccess] = useState(false);
 
 	const [isDuplicated, setIsDuplicated] = useState(false);
 	const [nickname, setNickname] = useState<string | null>(null);
@@ -31,10 +27,6 @@ export default function Page() {
 		privacy: false,
 		marketing: false,
 	});
-
-	const isValidNickname = nickname && !isDuplicated;
-	const isAllRequiredChecked = agreements.age && agreements.term && agreements.privacy;
-	const isButtonDisabled = !(isValidNickname && isAllRequiredChecked && teams);
 
 	const handleNicknameChange = (e) => {
 		setNickname(e.target.value);
@@ -61,38 +53,9 @@ export default function Page() {
 		}
 	};
 
-	const handleSignupButtonClick = async () => {
-		// 회원가입(정보 수정)
-		const updateUserInfoRequest: UpdateUserInfoRequest = {
-			nickname: nickname,
-			teams: teams,
-		};
-		const updateUserInfoResponse = await updateUserInfo(updateUserInfoRequest);
-
-		if (updateUserInfoResponse === 'DUPLICATED_NICKNAME') {
-			if (isDuplicated === false) setIsDuplicated(true); // 닉네임 중복
-		} else if (typeof updateUserInfoResponse === 'string') {
-			if (isDuplicated === true) setIsDuplicated(false); // 기타 오류
-			console.log(updateUserInfoResponse);
-		} else {
-			// 성공 시 약관 동의 api 호출
-			const privacyRequest: UpdatePrivacyRequest = {
-				privacyAgreedAt: agreements.privacy && new Date().toISOString().split('.')[0] + 'Z',
-				marketingAgreedAt: agreements.marketing ? new Date().toISOString().split('.')[0] + 'Z' : undefined,
-			};
-			const privacyResponse = await updatePrivacy(privacyRequest);
-
-			if (typeof privacyResponse === 'string') {
-				console.log(privacyResponse);
-			} else {
-				// 성공 시 재로그인
-				const redirectUrl = `${DOMAIN_URL || 'http://localhost:3000'}/api/auth/${provider}/callback`;
-				router.push(`${SERVER_URL}/oauth2/authorization/${provider}?state=${redirectUrl}`);
-			}
-		}
-	};
-
 	// 소셜 로그인을 통한 접근이 아닌 경우 홈으로 리디렉션
+	const [isValidAccess, setIsValidAccess] = useState(false);
+
 	useEffect(() => {
 		const fromLogin = getCookie('fromLogin');
 
@@ -112,7 +75,14 @@ export default function Page() {
 		};
 	}, [isValidAccess]);
 
-	const buttons = [{ text: '회원가입', onClick: handleSignupButtonClick, disabled: isButtonDisabled }];
+	const buttonSectionProps = {
+		provider,
+		nickname,
+		teams,
+		isDuplicated,
+		setIsDuplicated,
+		agreements,
+	};
 
 	return (
 		<div className="w-[21.5rem] m-auto flex flex-col items-center">
@@ -139,7 +109,7 @@ export default function Page() {
 					/>
 				))}
 
-				<BottomButton buttons={buttons} />
+				<ButtonSection {...buttonSectionProps} />
 			</div>
 		</div>
 	);
